refactor PreferSetView
